Add auth

+++ Kitbox/Classes/CupboardAngle.cs
@@ -9,7 +9,7 @@ namespace Kitbox.Classes
     class CupboardAngle : Specs
     {
         public string Color { get; set; }
-        public CupboardAngle(string color, float height, float width, float depth) : base(height, width, depth)
+        public CupboardAngle(string color, int height, int width, int depth) : base(height, width, depth)
         {
             this.Color = color;
         }
